Api versioning
* Release/0.0.1

* Add Patch handle And set delete status code

* Golang go mod pattern

* Use of Service

* refactor: use of service

* build: tests

* Add Handlers Tests

* test: handler delete

* test: handler get byFairID

* test: handler get byFairID

* test: handler post NewFair

* test: general adjustments

* test: handler put updateFair

* New handler fair by query

* Tests for repository

* test: add tests for repository

* test: add test for DeleteByID

* test: add test for Save

* test: add test for update

* Add Adrs

* feat: api versioning

* docs: documentation adjustments

* refactor: refactors
