+++ internal/handlers/v1/delete/fairByID_test.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"github.com/gin-gonic/gin"
 	"github.com/jeffersonto/feira-api/cmd/server/middleware"
-	"github.com/jeffersonto/feira-api/internal/handlers"
-	"github.com/jeffersonto/feira-api/internal/handlers/delete"
+	"github.com/jeffersonto/feira-api/internal/handlers/v1"
+	"github.com/jeffersonto/feira-api/internal/handlers/v1/delete"
 	"github.com/jeffersonto/feira-api/internal/service"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -70,10 +70,11 @@ func TestFairByID(t *testing.T) {
 			tt.warmUP(tt.expectedFairID)
 			router := gin.Default()
 			router.Use(middleware.ErrorHandle())
-			handler := handlers.NewHandler(service)
-			delete.NewFairByIDyHandler(handler, router)
+			routerGroupV1 := router.Group("/v1")
+			handler := v1.NewHandler(service, routerGroupV1)
+			delete.NewFairByIDyHandler(handler)
 			response := httptest.NewRecorder()
-			req, _ := http.NewRequest("DELETE", fmt.Sprintf("/feiras/%v", tt.pathParameter), nil)
+			req, _ := http.NewRequest("DELETE", fmt.Sprintf("/v1/feiras/%v", tt.pathParameter), nil)
 			router.ServeHTTP(response, req)
 			tt.expected(response)
 		})
