+++ internal/handlers/v1/put/updateFair_test.go
@@ -1,4 +1,4 @@
-package post_test
+package put_test
 
 import (
 	"bytes"
@@ -6,8 +6,8 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/jeffersonto/feira-api/cmd/server/middleware"
 	"github.com/jeffersonto/feira-api/internal/dto"
-	"github.com/jeffersonto/feira-api/internal/handlers"
-	"github.com/jeffersonto/feira-api/internal/handlers/post"
+	"github.com/jeffersonto/feira-api/internal/handlers/v1"
+	"github.com/jeffersonto/feira-api/internal/handlers/v1/put"
 	"github.com/jeffersonto/feira-api/internal/service"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -16,18 +16,20 @@ import (
 	"testing"
 )
 
-func TestNewFair(t *testing.T) {
+func TestUpdateFair(t *testing.T) {
 
 	var service *serviceMock
 
 	tests := []struct {
-		name     string
-		body     string
-		warmUP   func()
-		expected func(result *httptest.ResponseRecorder)
+		name          string
+		pathParameter string
+		body          string
+		warmUP        func()
+		expected      func(result *httptest.ResponseRecorder)
 	}{
 		{
-			name: "Should successfully get and return status code 204",
+			name:          "Should successfully get and return status code 200",
+			pathParameter: "1",
 			body: `{
 						"longitude": -46550164,
 						"latitude": -23558733,
@@ -48,15 +50,16 @@ func TestNewFair(t *testing.T) {
 				}`,
 			warmUP: func() {
 				service = new(serviceMock)
-				service.On("SaveFair", mock.Anything).Return(nil)
+				service.On("UpdateFairByID", mock.Anything, mock.Anything).Return(nil)
 			},
 			expected: func(result *httptest.ResponseRecorder) {
-				assert.Equal(t, 201, result.Code)
-				service.AssertNumberOfCalls(t, "SaveFair", 1)
+				assert.Equal(t, 200, result.Code)
+				service.AssertNumberOfCalls(t, "UpdateFairByID", 1)
 			},
 		},
 		{
-			name: "Should miss the longitude field and return status code 400",
+			name:          "Should miss the longitude field and return status code 400",
+			pathParameter: "1",
 			body: `{
 						"latitude": -23558733,
 						"setor_censitario": 355030885000091,
@@ -76,15 +79,16 @@ func TestNewFair(t *testing.T) {
 				}`,
 			warmUP: func() {
 				service = new(serviceMock)
-				service.On("SaveFair", mock.Anything).Return(nil)
+				service.On("UpdateFairByID", mock.Anything, mock.Anything).Return(nil)
 			},
 			expected: func(result *httptest.ResponseRecorder) {
 				assert.Equal(t, 400, result.Code)
-				service.AssertNumberOfCalls(t, "SaveFair", 0)
+				service.AssertNumberOfCalls(t, "UpdateFairByID", 0)
 			},
 		},
 		{
-			name: "Should execute the SaveFair Function, however receive an internal_server_error with status code 500",
+			name:          "Should not be able to convert the path parameter and return status code 400",
+			pathParameter: "A",
 			body: `{
 						"longitude": -46550164,
 						"latitude": -23558733,
@@ -105,11 +109,42 @@ func TestNewFair(t *testing.T) {
 				}`,
 			warmUP: func() {
 				service = new(serviceMock)
-				service.On("SaveFair", mock.Anything).Return(fmt.Errorf("internal_server_error"))
+				service.On("UpdateFairByID", mock.Anything, mock.Anything).Return(nil)
+			},
+			expected: func(result *httptest.ResponseRecorder) {
+				assert.Equal(t, 400, result.Code)
+				service.AssertNumberOfCalls(t, "UpdateFairByID", 0)
+			},
+		},
+		{
+			name:          "Should execute the SaveFair Function, however receive an internal_server_error with status code 500",
+			pathParameter: "1",
+			body: `{
+						"longitude": -46550164,
+						"latitude": -23558733,
+						"latitude": -23558733,
+						"setor_censitario": 355030885000091,
+						"area_ponderacao": 3550308005040,
+						"codigo_ibge": "87",
+						"distrito": "VILA FORMOSA",
+						"codigo_subprefeitura": 26,
+						"subprefeitura": "ARICANDUVA-FORMOSA-CARRAO",
+						"regiao5": "Leste",
+						"regiao8": "Leste 1",
+						"nome_feira": "VILA FORMOSA",
+						"registro": "4041-0",
+						"logradouro": "RUA MARAGOJIPE",
+						"numero": "S/N",
+						"bairro": "VL FORMOSA",
+						"referencia": "TV RUA PRETORIA"
+				}`,
+			warmUP: func() {
+				service = new(serviceMock)
+				service.On("UpdateFairByID", mock.Anything, mock.Anything).Return(fmt.Errorf("internal_server_error"))
 			},
 			expected: func(result *httptest.ResponseRecorder) {
 				assert.Equal(t, 500, result.Code)
-				service.AssertNumberOfCalls(t, "SaveFair", 1)
+				service.AssertNumberOfCalls(t, "UpdateFairByID", 1)
 			},
 		},
 	}
@@ -118,10 +153,11 @@ func TestNewFair(t *testing.T) {
 			tt.warmUP()
 			router := gin.Default()
 			router.Use(middleware.ErrorHandle())
-			handler := handlers.NewHandler(service)
-			post.NewFairHandler(handler, router)
+			routerGroupV1 := router.Group("/v1")
+			handler := v1.NewHandler(service, routerGroupV1)
+			put.NewUpdateHandler(handler)
 			response := httptest.NewRecorder()
-			req, _ := http.NewRequest("POST", "/feiras", bytes.NewBufferString(tt.body))
+			req, _ := http.NewRequest("PUT", fmt.Sprintf("/v1/feiras/%v", tt.pathParameter), bytes.NewBufferString(tt.body))
 			router.ServeHTTP(response, req)
 			tt.expected(response)
 		})
@@ -133,7 +169,7 @@ type serviceMock struct {
 	service.FairService
 }
 
-func (sm *serviceMock) SaveFair(newFair dto.Fair) error {
-	args := sm.Called(newFair)
+func (sm *serviceMock) UpdateFairByID(fairID int64, fairToBeUpdated dto.Fair) error {
+	args := sm.Called(fairID, fairToBeUpdated)
 	return args.Error(0)
 }
